more finessing of longColumn and bug-fix in datetime col.

+++ src/main/java/com/deathrayresearch/outlier/columns/LongColumn.java
@@ -82,7 +82,7 @@ public class LongColumn extends AbstractColumn implements LongMapUtils {
     data.add(i);
   }
 
-  public void set(int index, short value) {
+  public void set(int index, long value) {
     data.set(index, value);
   }
